requirements.txt
torch
transformers
qdrant-client
opencv-python
pillow
gradio
streamlit
segment-anything
opencv-python
supervision
git+https://github.com/facebookresearch/segment-anything-2.git